Refactor: Move sync backend to its own package
